backend/requirements.txt
fastapi
uvicorn[standard]
pymysql
pyjwt[crypto]
fastapi-cors
azure-identity
pyodbc
python-dotenv
